+++ src/derivative.h
@@ -2,7 +2,7 @@ class derivative {
   public:
     derivative(float loopTime);
     float update(float data);
-    float update(float data, unsigned long dt);
+    float update(float data, float dt);
 
   private:
     float vLoopTime;
